meta: introduce dev and prod build types
If ARGUS_BUILD_TYPE is set to "prod" (default configuration), the
generated image will be suitable for production usage (as hardened as
possible, login with root user disabled, etc).

If ARGUS_BUILD_TYPE is set to "dev", the generated image will be
suitable for development (root logins allowed, etc). This can be done
by setting the variable in the local.conf file.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>